refactor: split carts service
Create separate carts services based on functionality
